weather/requirements.txt
tensorflow>=2.8.0
pandas>=1.3.0
numpy>=1.19.2
scikit-learn>=0.24.0
requests>=2.25.0
joblib>=1.0.0
aiohttp>=3.8.1
python-dotenv>=0.19.0